Moved linked proposal to new file
